refactor show task page details blocks
