+++ src/app.scss
@@ -1,7 +1,7 @@
-// Override default variables
-$primary-500: #e91e63;
-$primary-600: #d81b60;
-$primary-700: #c2185b;
+@import "./styles/variables.scss";
+
+// Import Pico
+@import "@picocss/pico/scss/pico";
 
 // additional variables
 :root {
@@ -11,9 +11,6 @@ $primary-700: #c2185b;
     --important-color: #ff6f3c;
     --done-color: #42b883;
 }
-
-// Import Pico
-@import "@picocss/pico/scss/pico";
 
 // adjust lucid icon alignement
 .lucide-icon:not(.tsk-Cat) { position: relative; top:-2.5px; }
