+++ src/styles/variables.scss
@@ -1,0 +1,4 @@
+// Override default variables (pico.css)
+$primary-500: #e91e63;
+$primary-600: #d81b60;
+$primary-700: #c2185b;
